<feature> buff对属性值进行影响，伤害加成和减免逻辑

+++ Assets/TiedanSouls/Client/World/Generic/Model/DamageRecordModel.cs
@@ -12,7 +12,7 @@ namespace TiedanSouls.Client {
         public EntityIDArgs victim;           // 受害者
 
         public DamageType damageType;   // 伤害类型
-        public int damage;              // 伤害值
+        public float damage;              // 伤害值
 
         public override string ToString() {
             return $"伤害记录数据\n攻击者:{attacker}\n受害者:{victim}\n伤害类型:{damageType}\n伤害值:{damage}";
